config for campaign

+++ wifi_stats/programs/files_transfer.py
@@ -2,14 +2,10 @@
 Send files continuously to stations in order to create traffic in the network.
 """
 
-# Importation des librairies
 from datetime import datetime, timedelta
-from time import sleep
 import random
-from random import randint
 from threading import Thread, Event
 import os
-from typing import Dict
 import yaml
 from datetime import datetime
 import logging
@@ -48,9 +44,10 @@ class FilesSender(Thread):
             self.wait_first_time = True
         self.kwargs["transfer_number"] = self.curr_occurrences -1
         if not self.finished.is_set():
-            logger.info(f"Iteration: {self.curr_occurrences}")
-            if self.curr_occurrences <= self.args[0]["transfer_nb_per_step"]:
-                time.sleep(20)
+            initial_dead_time = self.args[0]["initial_dead_time_in_secs"]
+            if self.curr_occurrences < self.args[0]["transfer_nb_per_step"]:
+                logger.info(f"waiting {initial_dead_time} secs")
+                time.sleep(initial_dead_time)
             else:
                 self.function(*self.args, **self.kwargs)
             self.curr_occurrences += 1
@@ -95,23 +92,25 @@ def send_file_to_station(station, files_path: str, transfer_number: int):
         data_rate = initial_data_rate + (throughput_increment * int(transfer_number / transfer_nb_per_step))
         _file = get_iteration_file(files_path,transfer_number, transfer_nb_per_step)
 
-    # Log start of transfer
-    _log_line = f"Sending file {_file} to {station_ip} rate: {data_rate} kbps "
-    logger.info(_log_line)
+    if os.path.exists(_file):
+        # Log start of transfer
+        logger.info(f"Iteration: {transfer_number}")
+        _log_line = f"Sending file {_file} to {station_ip} rate: {data_rate} kbps "
+        logger.info(_log_line)
 
-    # Run SCP file transfer
-    scp_command = (
-        f"sshpass -p '{ssh_password}' scp -l {data_rate} {_file} {ssh_usr}@{station_ip}:files_transfer/file.txt"
-    )
-    os.system(scp_command)
-    # sleep(0.2)
-    logger.info("File sent")
-    # scp_command_rm = (
-    #     f"sshpass -p '{ssh_password}' ssh {ssh_usr}@{station_ip} rm -r /home/{ssh_usr}/files_transfer/*"
-    # )
-    # os.system(scp_command_rm)
-    # sleep(0.2)
-    # logger.info("File removed")
+        # Run SCP file transfer
+        scp_command = (
+            f"sshpass -p '{ssh_password}' scp -l {data_rate} {_file} {ssh_usr}@{station_ip}:files_transfer/file.txt"
+        )
+        os.system(scp_command)
+        # sleep(0.2)
+        logger.info("File sent")
+        # scp_command_rm = (
+        #     f"sshpass -p '{ssh_password}' ssh {ssh_usr}@{station_ip} rm -r /home/{ssh_usr}/files_transfer/*"
+        # )
+        # os.system(scp_command_rm)
+        # sleep(0.2)
+        # logger.info("File removed")
 
 
 def get_test_params(config_file: str):
